requirements.txt
Django
djangorestframework
PyJWT
# psycopg2==2.6.2
python-jenkins